refactor: Refactor the Activity component
