+++ src/activity-feed/Activity.jsx
@@ -1,34 +1,32 @@
 import React from 'react'
 import { find } from 'lodash'
-
 import PropTypes from 'prop-types'
-import cards from './cards'
 
-export default class Activity extends React.PureComponent {
-  static propTypes = {
-    activity: PropTypes.object.isRequired,
-    showDetails: PropTypes.bool,
-    filter: PropTypes.array,
-  }
+import activities from './activities'
 
-  static defaultProps = {
-    showDetails: false,
-    filter: [],
-  }
+function Activity({ activity, showDetails, filter }) {
+  const ActivityToRender = find(activities, (a) =>
+    a.canRender(activity, filter)
+  )
 
-  constructor(props) {
-    super(props)
-    const { activity, filter } = this.props
-
-    this.Card = find(cards, (c) => c.canRender(activity, filter))
-  }
-
-  render() {
-    const { activity, filter, showDetails } = this.props
-    const { Card } = this
-
-    return Card ? (
-      <Card activity={activity} filter={filter} showDetails={showDetails} />
-    ) : null
-  }
+  return ActivityToRender ? (
+    <ActivityToRender
+      activity={activity}
+      filter={filter}
+      showDetails={showDetails}
+    />
+  ) : null
 }
+
+Activity.propTypes = {
+  activity: PropTypes.object.isRequired,
+  showDetails: PropTypes.bool,
+  filter: PropTypes.array,
+}
+
+Activity.defaultProps = {
+  showDetails: false,
+  filter: [],
+}
+
+export default Activity
