+++ src/activity-feed/activities/CompaniesHouseCompany.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { get } from 'lodash'
-
 import PropTypes from 'prop-types'
+
 import {
   Card,
   CardDetails,
