+++ src/activity-feed/activities/InteractionUtils.js
