+++ src/activity-feed/activities/__tests__/InteractionUtils.test.js
