+++ src/activity-feed/activities/card/CardDetails.jsx
