+++ src/activity-feed/activities/card/CardDetailsList.jsx
