+++ src/activity-feed/activities/card/CardHeadingBlock.jsx
