+++ src/activity-feed/activities/card/CardUtils.jsx
