+++ src/activity-feed/activities/card/__tests__/CardUtils.test.js
