+++ src/activity-feed/activities/card/index.js
