+++ src/activity-feed/activities/card/item-renderers/BooleanItemRenderer.jsx
