+++ src/activity-feed/activities/card/item-renderers/ContactItemRenderer.jsx
